Extract OrderFormProvider to a different file
