+++ react/OrderFormContext.js
@@ -1,5 +1,9 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
+import { graphql, compose } from 'react-apollo'
+
+import orderFormQuery from './queries/orderFormQuery.gql'
+import addToCartMutation from './mutations/addToCartMutation.gql'
 
 const defaultState = {
   loading: true,
@@ -7,38 +11,6 @@ const defaultState = {
   refetch: () => {},
   updateOrderForm: () => {},
 }
-
-export function retrieveStateFromProps(props) {
-  if (!props.data.loading && !props.data.error) {
-    let orderFormContext = props.data
-    orderFormContext.updateOrderForm = props.updateOrderForm
-
-    return {
-      orderFormContext,
-    }
-  }
-
-  return {
-    orderFormContext: defaultState,
-  }
-}
-
-export const contextPropTypes = PropTypes.shape({
-  loading: PropTypes.bool.isRequired,
-  /* Function to refetch the orderForm query */
-  refetch: PropTypes.func.isRequired,
-  /* Function to update the orderForm */
-  updateOrderForm: PropTypes.func.isRequired,
-  /* Order form */
-  orderForm: PropTypes.shape({
-    /* Order form id */
-    orderFormId: PropTypes.string,
-    /* Total price of the order */
-    value: PropTypes.number,
-    /* Items in the mini cart */
-    items: PropTypes.arrayOf(PropTypes.object),
-  }),
-}).isRequired
 
 const { Consumer, Provider } = React.createContext(defaultState)
 
@@ -57,11 +29,59 @@ function orderFormConsumer(WrappedComponent) {
   }
 }
 
-export { Provider as OrderFormProvider }
+class ContextProvider extends Component {
+  static propTypes = {
+    children: PropTypes.element,
+  }
 
-export const orderFormProps = {
-  orderForm: PropTypes.object.isRequired,
-  refetch: PropTypes.func.isRequired,
+  state = {}
+
+  static getDerivedStateFromProps(props) {
+    if (!props.data.loading && !props.data.error) {
+      let orderFormContext = props.data
+      orderFormContext.updateOrderForm = props.updateOrderForm
+
+      return {
+        orderFormContext,
+      }
+    }
+
+    return {
+      orderFormContext: defaultState,
+    }
+  }
+
+  render() {
+    return <Provider value={this.state}>{this.props.children}</Provider>
+  }
 }
 
-export default { orderFormConsumer }
+const options = {
+  options: () => ({
+    ssr: false,
+  }),
+}
+
+const contextPropTypes = PropTypes.shape({
+  loading: PropTypes.bool.isRequired,
+  /* Function to refetch the orderForm query */
+  refetch: PropTypes.func.isRequired,
+  /* Function to update the orderForm */
+  updateOrderForm: PropTypes.func.isRequired,
+  /* Order form */
+  orderForm: PropTypes.shape({
+    /* Order form id */
+    orderFormId: PropTypes.string,
+    /* Total price of the order */
+    value: PropTypes.number,
+    /* Items in the mini cart */
+    items: PropTypes.arrayOf(PropTypes.object),
+  }),
+})
+
+export const OrderFormProvider = compose(
+  graphql(orderFormQuery, options),
+  graphql(addToCartMutation, { name: 'updateOrderForm' })
+)(ContextProvider)
+
+export default { orderFormConsumer, contextPropTypes }
