+++ react/StoreContextProvider.js
@@ -1,25 +1,15 @@
 import React, { Component } from 'react'
-import { graphql, compose } from 'react-apollo'
 import PropTypes from 'prop-types'
-import { OrderFormProvider, retrieveStateFromProps } from './OrderFormContext'
+import { OrderFormProvider } from './OrderFormContext'
 import { DataLayerProvider } from './components/withDataLayer'
-
-import orderFormQuery from './queries/orderFormQuery.gql'
-import addToCartMutation from './mutations/addToCartMutation.gql'
 
 class StoreContextProvider extends Component {
   static propTypes = {
     children: PropTypes.element,
   }
 
-  state = {}
-
   pushToDataLayer = obj => {
     window.dataLayer.push(obj)
-  }
-
-  static getDerivedStateFromProps(props) {
-    return retrieveStateFromProps(props)
   }
 
   render() {
@@ -32,7 +22,7 @@ class StoreContextProvider extends Component {
           pushToDataLayer: this.pushToDataLayer,
         }}
       >
-        <OrderFormProvider value={this.state}>
+        <OrderFormProvider>
           <div className="vtex-store__template">{this.props.children}</div>
         </OrderFormProvider>
       </DataLayerProvider>
@@ -40,13 +30,4 @@ class StoreContextProvider extends Component {
   }
 }
 
-const options = {
-  options: () => ({
-    ssr: false,
-  }),
-}
-
-export default compose(
-  graphql(orderFormQuery, options),
-  graphql(addToCartMutation, { name: 'updateOrderForm' })
-)(StoreContextProvider)
+export default StoreContextProvider
